+++ aws_profile_manager/profile_remove/__init__.py
@@ -1,3 +1,0 @@
-from .remove import Remove
-
-__all__ = ('Remove',)